// src/api/generated/types.gen.ts
export type AddressDto = {
  label?: string;
  line1: string;
  line2?: string;
  city: string;
  state?: string;
  postalCode: string;
  country: string;
};

export type UserResponseDto = {
  /**
   * Firebase UID
   */
  id: string;
  email: string;
  role: 'CUSTOMER' | 'ADMIN';
  addresses: Array<AddressDto>;
  createdAt: string;
  updatedAt: string;
};